Conflicts:
	framework/src/ppl/delite/framework/a-ops/DeliteFileReader.scala
	framework/src/ppl/delite/framework/a-ops/DeliteOps.scala
	framework/src/ppl/delite/framework/codegen/delite/generators/DeliteGenTaskGraph.scala
	framework/src/ppl/delite/framework/datastructures/DeliteArray.scala
	framework/src/ppl/delite/framework/datastructures/DeliteArrayBuffer.scala
	framework/src/ppl/delite/framework/transform/MultiLoopSoA.scala
	runtime/src/ppl/delite/runtime/codegen/kernels/scala/ScalaMultiLoopGenerator.scala
	runtime/src/static/cpp/cppDeliteArray.h
	runtime/src/static/cpp/cppFileStream.h

+++ runtime/src/ppl/delite/runtime/codegen/CodeCache.scala
@@ -143,24 +143,21 @@ trait CodeCache {
     allMatch
   }
 
-  protected def copyDirectory(source: Directory, destination: Directory) {
-    if (destination exists)
-      destination.deleteRecursively()
+  protected def copyDirectory(source: Directory, destination: Directory, mergeDestination: Boolean = false) {
+    if (!mergeDestination) destination.deleteRecursively()
     destination.createDirectory()
 
-    val base = destination.toString + File.separator
-
     for (dir <- source.dirs)
-      copyDirectory(dir, Directory(Path(base + dir.name)))
+      copyDirectory(dir, destination / Directory(dir.name), mergeDestination)
 
     for (file <- source.files)
-      file.copyTo(Path(base + file.name))
+      file.copyTo(destination / file.name)
   }
 
   def printSources() {
     for (i <- 0 until sourceBuffer.length) {
       print(sourceBuffer(i))
-      print("\n /*********/ \n \n")
+      println("\n /*********/ \n")
     }
   }
